+++ code/framework/types/String.hpp
@@ -14,8 +14,12 @@ class String : public Data {
 		typedef const SMARTPOINTER<String> c_ptr;
 
 	public:
+		// default constructor
+		String() {};
 		// constructor
 		String(std::string s) : str_(s) {};
+		// copy constructor
+		String(const String& s) : str_(s.str_) {};
 		// destructor
 		~String(void){};